Homepage refactor
